Added Reconnect code to master and botinstances, so bots will come back when they lagout.
chancged amysql to have a 5 sec timeout when it .joins the db threead, i suspect this is what is keeping master from dying properly when i shutdown
fixed bug in startbot that was spawning mulitple bots of the same name...  the instance dit has  lowercase names in it. when i checked if that name was running i forgot to lowercase it

added some comments to SubspaceBot.py,
in  event scoreupdate. added event.old_flag_XXX value